requirements.txt
# 核心依赖
Pillow
gradio

# 数据处理
numpy

# 开发工具
python-dotenv
requests

# API服务
tiktoken
openai

# UI
matplotlib

# 其他
solara